+++ packages/frontend/src/os.ts
@@ -24,7 +24,7 @@ import MkContextMenu from '@/components/MkContextMenu.vue';
 import { MenuItem } from '@/types/menu.js';
 import copyToClipboard from '@/scripts/copy-to-clipboard.js';
 import { showMovedDialog } from '@/scripts/show-moved-dialog.js';
-import { getHTMLElementOrNull } from '@/scripts/tms/get-or-null.js';
+import { getHTMLElementOrNull } from '@/scripts/get-or-null.js';
 
 export const openingWindowsCount = ref(0);
